fix: fresh-eyes review of this session's code — softcap dtype parity + cleanups
Careful re-read of the code written/modified this session. Three findings fixed:

1. gpt_synaptic.py (real parity divergence): the vg9.1 softcap change cast logits to
   float32 UNCONDITIONALLY, but the vanilla GPT head floats ONLY on the loss path and
   returns inference logits in the lm_head dtype. So the synaptic INFERENCE return dtype
   changed (float32 vs lm_head dtype, e.g. bf16 on GPU) — a subtle divergence that could
   surprise downstream consumers. Now matches GPT exactly: softcap on both paths, but
   the inference path returns lm_head-dtype logits and only the loss path casts to fp32.

2. tests/_bio_testkit.py (no-op + misleading): set_seed set PYTHONHASHSEED via
   os.environ.setdefault, which has no effect on the running process (it only matters at
   interpreter startup) and the comment wrongly implied it made hash ordering
   reproducible. Removed.

3. tests/conftest.py: dropped an unnecessary `# noqa: E402` (no E402 was triggered);
   replaced with a plain explanatory comment.

Verified: full non-slow suite 90 passed / 4 skipped; ruff + ty clean.

+++ bio_inspired_nanochat/gpt_synaptic.py
@@ -277,12 +277,13 @@ class GPTSynaptic(nn.Module):
             kv_cache.presyn_state = presyn_states
 
         logits = self.lm_head(x.to(dtype=self.lm_head.weight.dtype))
-        # Logit softcap (parity with vanilla GPT) — applied on BOTH the inference
-        # return and the loss path so logits are bounded everywhere. See vg9.1.
+        # Logit softcap (parity with vanilla GPT) — bound logits on BOTH the inference
+        # return and the loss path. See vg9.1. We match GPT exactly: the inference
+        # path returns softcapped logits in the lm_head dtype (no float cast), and the
+        # loss path casts to fp32 before cross-entropy.
         sc = self.config.logit_softcap
         if sc and sc > 0.0:
             logits = sc * torch.tanh(logits / sc)
-        logits = logits.float()  # fp32 logits (parity with GPT)
         if targets is None:
             return logits, None
         aux = sum(
@@ -291,6 +292,7 @@ class GPTSynaptic(nn.Module):
                 for b in self.h
             )
         )
+        logits = logits.float()  # fp32 logits for the loss (parity with GPT)
         logits_flat = logits.reshape(-1, logits.size(-1))
         targets_flat = targets.reshape(-1)
         ce = F.cross_entropy(
